Extract TaxonomySelect to separate component
